feat(notification): DTO 패키지 구조 정리 및 클래스명 변경 (#174)
