+++ src/main/java/com/project200/undabang/admin/entity/dto/CommonErrorData.java
@@ -1,4 +1,4 @@
-package com.project200.undabang.admin.component.dto;
+package com.project200.undabang.admin.entity.dto;
 
 import lombok.Builder;
 import lombok.Getter;
