+++ src/main/java/com/project200/undabang/admin/entity/dto/context/AddScoreErrorData.java
@@ -1,4 +1,4 @@
-package com.project200.undabang.admin.component.dto.context;
+package com.project200.undabang.admin.entity.dto.context;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import java.util.UUID;
 
 @Getter
 @Builder
-public class UserErrorContext {
+public class AddScoreErrorData {
     private final String requestUri; // 요청 URI
     private final String httpMethod; // HTTP METHOD
     private final UUID userIdentifier; // 유저 식별자 UUID
