+++ src/main/java/com/project200/undabang/admin/entity/dto/context/BatchErrorData.java
@@ -1,11 +1,11 @@
-package com.project200.undabang.admin.component.dto.context;
+package com.project200.undabang.admin.entity.dto.context;
 
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-public class BatchErrorContext {
+public class BatchErrorData {
     private String jobName;
     private String jobParameters;
     private String status;
